# BASS_Env/openapi.yaml
openapi: 3.0.3
info:
  title: BinderLabs API Security Simulator
  description: Binderlabs API Security Simulator (BASS-Env) is a vulnerable API environment that is susceptible to the OWASP Top 10 API Security Risk – 2023. The environment was deliberately designed with API flaws and as a practice ground for cybersecurity professionals who wishes to hands-on their API hacking skills and enhance their knowledge in API Security testing domain.
  version: 1.0.0
  contact: {}
servers:
- url: http://{{host}}
- url: '{{host}}'
paths:
...
  /bass/api6/marketplace/premium/purchase:
    post:
      tags:
      - API6:Unrestricted Access to Sensitive Business Flows
      summary: Purchase
      description: Purchase
      operationId: purchase
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                purchases:
                  type: array
                  items:
                    type: object
                    properties:
                      item_id:
                        type: number
                        example: 1
                      quantity:
                        type: number
                        example: 2
                  example:
                  - item_id: 1
                    quantity: 2
                  - item_id: 2
                    quantity: 1
                uuid:
                  type: string
                  example: 87faf56f-5be6-4932-927b-16559d6755bd
            examples:
              Purchase:
                value:
                  purchases:
                  - item_id: 1
                    quantity: 2
                  - item_id: 2
                    quantity: 1
                  uuid: 87faf56f-5be6-4932-927b-16559d6755bd
      responses:
        '200':
          description: ''
...
components:
  securitySchemes: {}
...